pyproject.toml
[project]
name = "gans-data-diet"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "jupyter>=1.1.1",
    "matplotlib>=3.10.0",
    "numpy>=2.2.3",
    "pandas>=2.2.3",
    "scipy>=1.15.2",
    "seaborn>=0.13.2",
    "torchinfo>=1.8.0",
    "torchvision>=0.21.0",
    "tqdm>=4.67.1",
]

[dependency-groups]
dev = [
    "gans-data-diet",
]

[tool.uv.sources]
gans-data-diet = { workspace = true }
